+++ Source/Iwentys.EntityManager.WebApi/Controllers/StudentProfiles/Queries/GetStudentProfilesByCourseId.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using AutoMapper.QueryableExtensions;
 using Iwentys.EntityManager.DataAccess;
 using Iwentys.EntityManager.WebApiDtos;
 using MediatR;
@@ -6,9 +7,9 @@ using Microsoft.EntityFrameworkCore;
 
 namespace Iwentys.EntityManager.WebApi;
 
-public static class GetStudentProfileByGroupId
+public static class GetStudentProfilesByCourseId
 {
-    public record Query(int GroupId) : IRequest<Response>;
+    public record Query(int CourseId) : IRequest<Response>;
     public record Response(IReadOnlyCollection<StudentInfoDto> Students);
 
     public class Handler : IRequestHandler<Query, Response>
@@ -24,11 +25,13 @@ public static class GetStudentProfileByGroupId
 
         public async Task<Response> Handle(Query request, CancellationToken cancellationToken)
         {
-            List<StudentInfoDto> result = await _mapper
-                .ProjectTo<StudentInfoDto>(_context.Students)
-                .Where(s => s.GroupId == request.GroupId)
+            List<StudentInfoDto> result = await _context
+                .StudyGroups
+                .Where(g => g.StudyCourseId == request.CourseId)
+                .SelectMany(g => g.Students)
+                .ProjectTo<StudentInfoDto>(_mapper.ConfigurationProvider)
                 .ToListAsync(cancellationToken: cancellationToken);
-
+            
             return new Response(result);
         }
     }
