+++ Source/Iwentys.EntityManager.WebApi/Controllers/StudentProfiles/StudentProfileController.cs
@@ -22,43 +22,38 @@ public class StudentProfileController : ControllerBase
         return Ok(response.Students);
     }
 
+    [HttpGet(nameof(GetStudentProfileByGithubUsername))]
+    public async Task<ActionResult<StudentInfoDto>> GetStudentProfileByGithubUsername(string githubUsername)
+    {
+        GetStudentProfileByGithubUsername.Response response = await _mediator.Send(new GetStudentProfileByGithubUsername.Query(githubUsername));
+        var result = response?.Student;
+
+        return result is not null ? Ok(result) : NotFound();
+    }
+
     [HttpGet(nameof(GetStudentProfileById))]
     public async Task<ActionResult<StudentInfoDto>> GetStudentProfileById(int id)
     {
         GetStudentProfileById.Response response = await _mediator.Send(new GetStudentProfileById.Query(id));
+        var result = response?.Student;
 
-        if (response?.Student is null) return NotFound();
-
-        return Ok(response.Student);
+        return result is not null ? Ok(result) : NotFound();
     }
 
     [HttpPost(nameof(GetStudentProfilesByIdList))]
     public async Task<ActionResult<IReadOnlyCollection<StudyGroupProfileResponseDto>>> GetStudentProfilesByIdList(
         List<int> studentIdList)
     {
-        GetStudentProfilesByIdList.Response response
-            = await _mediator.Send(new GetStudentProfilesByIdList.Query(studentIdList));
+        GetStudentProfilesByIdList.Response response = await _mediator.Send(new GetStudentProfilesByIdList.Query(studentIdList));
 
         return Ok(response.Students);
-    }
-
-    [HttpGet(nameof(GetStudentProfileByGithubUsername))]
-    public async Task<ActionResult<StudentInfoDto>> GetStudentProfileByGithubUsername(string githubUsername)
-    {
-        GetStudentProfileByGithubUsername.Response response
-            = await _mediator.Send(new GetStudentProfileByGithubUsername.Query(githubUsername));
-
-        if (response?.Student is null) return NotFound();
-
-        return Ok(response.Student);
     }
 
     [HttpPost(nameof(GetStudentProfilesByGithubUsernamesList))]
     public async Task<ActionResult<IReadOnlyCollection<StudentInfoDto>>> GetStudentProfilesByGithubUsernamesList(
         List<string> githubUsernamesList)
     {
-        GetStudentProfilesByGithubUsernamesList.Response response
-            = await _mediator.Send(new GetStudentProfilesByGithubUsernamesList.Query(githubUsernamesList));
+        GetStudentProfilesByGithubUsernamesList.Response response = await _mediator.Send(new GetStudentProfilesByGithubUsernamesList.Query(githubUsernamesList));
 
         return Ok(response.Students);
     }
@@ -66,8 +61,7 @@ public class StudentProfileController : ControllerBase
     [HttpGet(nameof(GetStudentProfilesByGroupId))]
     public async Task<ActionResult<IReadOnlyCollection<StudentInfoDto>>> GetStudentProfilesByGroupId(int groupId)
     {
-        GetStudentProfileByGroupId.Response response
-            = await _mediator.Send(new GetStudentProfileByGroupId.Query(groupId));
+        GetStudentProfilesByGroupId.Response response = await _mediator.Send(new GetStudentProfilesByGroupId.Query(groupId));
 
         return Ok(response.Students);
     }
@@ -75,8 +69,7 @@ public class StudentProfileController : ControllerBase
     [HttpGet(nameof(GetStudentProfilesByCourseId))]
     public async Task<ActionResult<IReadOnlyCollection<StudentInfoDto>>> GetStudentProfilesByCourseId(int courseId)
     {
-        GetStudentProfileByCourseId.Response response
-            = await _mediator.Send(new GetStudentProfileByCourseId.Query(courseId));
+        GetStudentProfilesByCourseId.Response response = await _mediator.Send(new GetStudentProfilesByCourseId.Query(courseId));
 
         return Ok(response.Students);
     }
@@ -85,8 +78,7 @@ public class StudentProfileController : ControllerBase
     public async Task<ActionResult<IReadOnlyCollection<StudentInfoDto>>> GetStudentProfilesByCredentials(
         string credentials)
     {
-        GetStudentProfilesByCredentials.Response response
-            = await _mediator.Send(new GetStudentProfilesByCredentials.Query(credentials));
+        GetStudentProfilesByCredentials.Response response = await _mediator.Send(new GetStudentProfilesByCredentials.Query(credentials));
 
         return Ok(response.Students);
     }
@@ -95,8 +87,7 @@ public class StudentProfileController : ControllerBase
     [HttpGet(nameof(GetStudentProfilesBySubjectId))]
     public async Task<ActionResult<IReadOnlyCollection<StudentInfoDto>>> GetStudentProfilesBySubjectId(int subjectId)
     {
-        GetStudentProfilesBySubjectId.Response response
-            = await _mediator.Send(new GetStudentProfilesBySubjectId.Query(subjectId));
+        GetStudentProfilesBySubjectId.Response response = await _mediator.Send(new GetStudentProfilesBySubjectId.Query(subjectId));
 
         return Ok(response.Students);
     }
